First commit
This files contain a color guessing game for the player. The file  play_part1.html runs into a loop until the player guesses the right color.Once the player guesses the right color it shows an alert message with the description of number of guesses it took for the player to guess the right color and sets the background color into that color. 
The play_part2.html file is almost same as play_part1.html.But, it contains additional features like if the player's input color does not belong to the array as given in the code it shows a warning message and continues with the loop. In case player's input color is alphabetically higher/lower than that of target color for that also it shows the respective alert message.